+++ CHANGELOG.rst
@@ -4,7 +4,8 @@ CHANGELOG
 5.3.69 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Try not to reserve connection when possible
+  [vangheem]
 
 
 5.3.68 (2021-04-16)